requirement.txt
pyserial
minimialmobdus
pymodbus
numpy